requirements.txt
fastapi
uvicorn[standard]
jinja2
pydantic
httpx
python-multipart
openai
python-docx
Pillow
sqlmodel
python-jose
passlib[bcrypt]
bcrypt
opencv-python-headless
easyocr
opentelemetry-instrumentation-openai
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-sqlalchemy
opentelemetry-instrumentation-logging